additional two compilation fixes

+++ src/main/hello_world_main.c
@@ -37,10 +37,10 @@ void app_main(void)
         return;
     }
 
-    printf("%uMB %s flash\n", flash_size / (1024 * 1024),
+    printf("%uMB %s flash\n", (unsigned int)(flash_size / (1024 * 1024)),
            (chip_info.features & CHIP_FEATURE_EMB_FLASH) ? "embedded" : "external");
 
-    printf("Minimum free heap size: %d bytes\n", esp_get_minimum_free_heap_size());
+    printf("Minimum free heap size: %d bytes\n", (int)esp_get_minimum_free_heap_size());
 
     for (int i = 10; i >= 0; i--) {
         printf("Restarting in %d seconds...\n", i);
